Added a %cwd var to catcierge_execute.
So that the commands run on different events, such as --save_imgs_cmd can get the absolute path that are relative to the grabber program. The commands will otherwise try to open the file relative to their own path and fail. There is no easy way to get the absolute path for non-existant files. the realpath(..) function only returns the absolute path of existing files. So simply passing the absolute paths is not possible in all cases.