Wait until the frame is cleared on a successful match.
After we have performed a successful match over the set of images, before we would simply start a timer before we tried to match again. However, from some real world testing it turns out this isn't enough.

When it's raining or snowing the cat likes to go out through the corridor where the match is made. This will result in "OK" match, and the timer starts... However, the cat doesn't feel like getting wet, so it will just sit there and look outsid, still in the corridor. So after the timer ends, the full frame will be occupied by the cats body and we will get a FAILED match.

This new approach instead waits until the frame is cleared after a successful match before the timer starts. This probably removes the need for the timer, but I'll still keep it but set the default to 0 seconds.