Add a limit for consecutive lockouts right after each other.
This is a safety precausion. For instance, when the backlight failed for the hardware, it would continously see a black picture and the door would stay locked (and I would be spammed via mail/twitter) about it in the middle of the night.

Instead, you can now specify a max count before the program forces an open of the door and simply quits.

Also added a "lockout dummy". Everything in the program works as normal, except that hte actual lockout action is never performed.