Rewrite frame obstructed function.
First of all, invert the image when doing the threshold so that cvSum will return 0 if there is nothing in the image, instead of having to calculate some expected sum of white pixels. This is a lot more common sense.

Also, don't count a 1 pixel obstruction like a real obstruction... In the summer it turns out spiders like to hang around in front of the camera and trigger it in the middle of the night.