Added template_path variable for output.
When generating the templates we save the generated paths.
By doing this we can now refer to the path of a generated template inside the
template itself (not very useful), or it can be used when passing
variables to catcierge_execute (to be added)...

This way, instead of passing loads of data via the command line in
catcierge_execute, we can instead simply create a template with the
desired input to the external program, and then simply pass the path
to that generated template to it.