Fix lockout timer bug.
When lockout timer method 2 (Wait for no obstruction, then timer) was used.

It would only work once after program start, since the lockout timer never was reset.
Instead, this would result in an immediate new match attempt after a failed one, which
usually meant the cat could get in with a prey.

Made it so that all the timer tests for this are now run twice in a row to catch this.

The fix was to reset the lockout timer after the end of the Lockout, so that it would
not be active at the next use.