Enable setting a different base time.
To make it possible to import all old images into RethinkDB I added the possiblity to change the base time for all time output in the program. This way I can parse the time and date saved in the old filename, generate images/json for all of them with this correct time (instead of the current time).