Bugfixes
When looking for the consecutive lockout bug it turned out fixing it triggered another bug which I also fixed.

Consecutive lockout bug => When we were supposed to stop running, instead we directly went from that state into locking, which is the complete opposite from what we want.

Now only run a set state if the actual state machine is running.