Added --auto_roi option.
After the initial --startup_delay, this setting will detect where in the image the backlight is, and use that region automatically as the region of interest for the is_obstructed check.

This was added because my crappy camera mount made the camera shift a bit too much and the obstruct detection would detect some stuff outside of the back light area, and think something was in the frame..... constantly... Which hits the max consecutive lockout count and shuts down.

Also this will be a good way to find out after startup if the back light is on or not (it might have failed). And quit right away, instead of spamming until consecutive lockout max has been hit (if it's even set that is).