Change how base_time is parsed
Also, don't leak memory when cargo_parse is called twice.

Instead of setting the actual base_time diff when doing the parsing, we instead simply save
the time diff and leave it up to the program to do it later. Since this setting makes
no sense at all in catcierge_grabber for instance, it makes no sense to use it there,
but in catcierge_fsm_tester it can be used.